# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/multisigs": {
            "get": {
                "description": "Get multisigs list with pagination, search, and sort",
                "tags": [
                    "Multisigs"
                ],
                "summary": "List Multisigs",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Page size",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Search keyword",
                        "name": "search",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Sort field and order, e.g., 'id desc'",
                        "name": "sort",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/main.Response"
                        }
                    }
                }
            },
            "post": {
                "description": "Create a new multisig",
                "tags": [
                    "Multisigs"
                ],
                "summary": "Create Multisig",
                "parameters": [
                    {
                        "description": "Multisig object",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/main.Multisig"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/main.Response"
                        }
                    }
                }
            }
        },
        "/multisigs/{multisig_address}": {
            "get": {
                "description": "Get a single multisig by address",
                "tags": [
                    "Multisigs"
                ],
                "summary": "Get Multisig",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Multisig Address",
                        "name": "multisig_address",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/main.Response"
                        }
                    }
                }
            },
            "put": {
                "description": "Update an existing multisig by address",
                "tags": [
                    "Multisigs"
                ],
                "summary": "Update Multisig",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Multisig Address",
                        "name": "multisig_address",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Updated multisig object",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/main.Multisig"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/main.Response"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete a multisig and all its related Vaults, Members, and Spends",
                "tags": [
                    "Multisigs"
                ],
                "summary": "Delete Multisig",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Multisig Address",
                        "name": "multisig_address",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/main.Response"
                        }
                    }
                }
            }
        },
        "/multisigs/{multisig_address}/members": {
            "get": {
                "description": "Get members list of a multisig with pagination, search, sort",
                "tags": [
                    "Members"
                ],
                "summary": "List Members",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Multisig Address",
                        "name": "multisig_address",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Page size",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Search keyword",
                        "name": "search",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Sort field and order, e.g., 'id desc'",
                        "name": "sort",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/main.Response"
                        }
                    }
                }
            },
            "post": {
                "description": "Create a new member under a multisig",
                "tags": [
                    "Members"
                ],
                "summary": "Create Member",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Multisig Address",
                        "name": "multisig_address",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Member object",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/main.Member"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/main.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "main.Member": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "member_address": {
                    "type": "string"
                },
                "multisig_address": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "main.Multisig": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "multisig_address": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "main.Response": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
...
    }
}